Rearranged file location of nodes and cloudformation command
