+++ src/explorer/awsExplorer.ts
@@ -13,8 +13,6 @@ import { deployLambda } from '../lambda/commands/deployLambda'
 import { getLambdaConfig } from '../lambda/commands/getLambdaConfig'
 import { invokeLambda } from '../lambda/commands/invokeLambda'
 import { newLambda } from '../lambda/commands/newLambda'
-import { FunctionNode } from '../lambda/explorer/functionNode'
-import { RegionNode } from '../lambda/explorer/regionNode'
 import { DefaultLambdaPolicyProvider, LambdaPolicyView } from '../lambda/lambdaPolicy'
 import * as utils from '../lambda/utils'
 import { AwsContext } from '../shared/awsContext'
@@ -24,6 +22,8 @@ import { ResourceFetcher } from '../shared/resourceFetcher'
 import { AWSCommandTreeNode } from '../shared/treeview/awsCommandTreeNode'
 import { AWSTreeNodeBase } from '../shared/treeview/awsTreeNodeBase'
 import { RefreshableAwsTreeProvider } from '../shared/treeview/refreshableAwsTreeProvider'
+import { FunctionNode } from './nodes/functionNode'
+import { RegionNode } from './nodes/regionNode'
 
 export class AwsExplorer implements vscode.TreeDataProvider<AWSTreeNodeBase>, RefreshableAwsTreeProvider {
     public static readonly VIEW_PROVIDER_ID: string = 'aws-explorer'
