+++ src/lambda/commands/deployLambda.ts
@@ -6,7 +6,7 @@
 'use strict'
 
 import * as vscode from 'vscode'
-import { FunctionNode } from '../explorer/functionNode'
+import { FunctionNode } from '../../explorer/nodes/functionNode'
 
 export async function deployLambda(element?: FunctionNode) {
     try {
