+++ src/test/explorer/awsExplorer.test.ts
@@ -7,7 +7,7 @@ import * as assert from 'assert'
 import * as AWS from 'aws-sdk'
 import * as vscode from 'vscode'
 import { AwsExplorer } from '../../explorer/awsExplorer'
-import { RegionNode } from '../../lambda/explorer/regionNode'
+import { RegionNode } from '../../explorer/nodes/regionNode'
 import { AwsContext, ContextChangeEventsArgs } from '../../shared/awsContext'
 import { AwsContextTreeCollection } from '../../shared/awsContextTreeCollection'
 import { RegionInfo } from '../../shared/regions/regionInfo'
